# api-reference/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "保险产品API",
    "version": "1.0.0"
  },
  "paths": {
    "/api/products": {
      "get": {
        "summary": "获取产品列表",
        "parameters": [
          {
            "name": "category",
            "in": "query",
            "description": "产品类别（如：团意险/雇主险）",
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "summary": "更新产品信息",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productCode": {
                    "type": "string",
                    "description": "产品代码"
                  },
                  "productName": {
                    "type": "string",
                    "description": "产品名称"
                  },
                  "coverage": {
                    "type": "object",
                    "description": "保障责任"
                  },
                  "rules": {
                    "type": "object",
                    "description": "承保规则"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{productCode}": {
      "get": {
        "summary": "获取产品详情",
        "parameters": [
          {
            "name": "productCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "产品代码"
          }
        ],
        "responses": {
          "200": {
            "description": "成功返回产品信息"
          }
        }
      }
    },
    "/api/policies": {
      "post": {
        "summary": "创建保单",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productCode": {
                    "type": "string",
                    "description": "产品代码"
                  },
                  "insuredList": {
                    "type": "array",
                    "description": "被保险人列表"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}